# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "FlexPos.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5088/"
    }
  ],
  "paths": {
...
    "/api/comisiones/reglas": {
      "get": {
        "tags": [
          "Comisiones"
        ],
        "parameters": [
          {
            "name": "empleadoId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "servicioId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pagina",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "tamanoPagina",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaReglasComisionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaReglasComisionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaReglasComisionDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Comisiones"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearReglaComisionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearReglaComisionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CrearReglaComisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReglaComisionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReglaComisionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReglaComisionDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/comisiones/reglas/{id}": {
      "put": {
        "tags": [
          "Comisiones"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarReglaComisionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarReglaComisionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarReglaComisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReglaComisionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReglaComisionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReglaComisionDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/comisiones/reglas/{id}/estado": {
      "patch": {
        "tags": [
          "Comisiones"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstadoReglaComisionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EstadoReglaComisionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EstadoReglaComisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReglaComisionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReglaComisionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReglaComisionDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/comisiones": {
      "get": {
        "tags": [
          "Comisiones"
        ],
        "parameters": [
          {
            "name": "empleadoId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "estado",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tipoMovimiento",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "desdeUtc",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "hastaUtc",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "pagina",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "tamanoPagina",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaComisionesDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaComisionesDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaComisionesDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/comisiones/liquidaciones": {
      "get": {
        "tags": [
          "Comisiones"
        ],
        "parameters": [
          {
            "name": "empleadoId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "desdeUtc",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "hastaUtc",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "pagina",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "tamanoPagina",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaLiquidacionesComisionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaLiquidacionesComisionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaLiquidacionesComisionDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Comisiones"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LiquidarComisionesRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LiquidarComisionesRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LiquidarComisionesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LiquidacionComisionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiquidacionComisionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiquidacionComisionDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/comisiones/liquidaciones/{id}": {
      "get": {
        "tags": [
          "Comisiones"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LiquidacionComisionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiquidacionComisionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiquidacionComisionDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActualizarReglaComisionRequest": {
        "required": [
          "tipo",
          "valor"
        ],
        "type": "object",
        "properties": {
          "tipo": {
            "type": [
              "null",
              "string"
            ]
          },
          "valor": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "ComisionDto": {
        "required": [
          "id",
          "empleadoId",
          "servicioId",
          "ventaId",
          "detalleVentaId",
          "tipoMovimiento",
          "tipoTarifa",
          "valorTarifa",
          "baseCalculo",
          "cantidad",
          "importe",
          "codigoMoneda",
          "estado",
          "comisionOriginalId",
          "devolucionVentaId",
          "liquidacionComisionId",
          "fechaCreacionUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "empleadoId": {
            "type": "string",
            "format": "uuid"
          },
          "servicioId": {
            "type": "string",
            "format": "uuid"
          },
          "ventaId": {
            "type": "string",
            "format": "uuid"
          },
          "detalleVentaId": {
            "type": "string",
            "format": "uuid"
          },
          "tipoMovimiento": {
            "type": "string"
          },
          "tipoTarifa": {
            "type": "string"
          },
          "valorTarifa": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "baseCalculo": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "cantidad": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "importe": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "codigoMoneda": {
            "type": "string"
          },
          "estado": {
            "type": "string"
          },
          "comisionOriginalId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "devolucionVentaId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "liquidacionComisionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "fechaCreacionUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CrearReglaComisionRequest": {
        "required": [
          "empleadoId",
          "servicioId",
          "tipo",
          "valor"
        ],
        "type": "object",
        "properties": {
          "empleadoId": {
            "type": "string",
            "format": "uuid"
          },
          "servicioId": {
            "type": "string",
            "format": "uuid"
          },
          "tipo": {
            "type": [
              "null",
              "string"
            ]
          },
          "valor": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "EstadoReglaComisionRequest": {
        "required": [
          "activa"
        ],
        "type": "object",
        "properties": {
          "activa": {
            "type": "boolean"
          }
        }
      },
...
      "LiquidacionComisionDto": {
        "required": [
          "id",
          "empleadoId",
          "metodoPagoId",
          "metodoPagoNombre",
          "esEfectivo",
          "importe",
          "codigoMoneda",
          "referencia",
          "cajaId",
          "fechaUtc",
          "comisionIds"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "empleadoId": {
            "type": "string",
            "format": "uuid"
          },
          "metodoPagoId": {
            "type": "string",
            "format": "uuid"
          },
          "metodoPagoNombre": {
            "type": "string"
          },
          "esEfectivo": {
            "type": "boolean"
          },
          "importe": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "codigoMoneda": {
            "type": "string"
          },
          "referencia": {
            "type": [
              "null",
              "string"
            ]
          },
          "cajaId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "fechaUtc": {
            "type": "string",
            "format": "date-time"
          },
          "comisionIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "LiquidarComisionesRequest": {
        "required": [
          "comisionIds",
          "metodoPagoId",
          "referencia"
        ],
        "type": "object",
        "properties": {
          "comisionIds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "metodoPagoId": {
            "type": "string",
            "format": "uuid"
          },
          "referencia": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PaginaComisionesDto": {
        "required": [
          "elementos",
          "pagina",
          "tamanoPagina",
          "totalElementos"
        ],
        "type": "object",
        "properties": {
          "elementos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComisionDto"
            }
          },
          "pagina": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "tamanoPagina": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalElementos": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "PaginaLiquidacionesComisionDto": {
        "required": [
          "elementos",
          "pagina",
          "tamanoPagina",
          "totalElementos"
        ],
        "type": "object",
        "properties": {
          "elementos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LiquidacionComisionDto"
            }
          },
          "pagina": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "tamanoPagina": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalElementos": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "PaginaReglasComisionDto": {
        "required": [
          "elementos",
          "pagina",
          "tamanoPagina",
          "totalElementos"
        ],
        "type": "object",
        "properties": {
          "elementos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReglaComisionDto"
            }
          },
          "pagina": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "tamanoPagina": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalElementos": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ReglaComisionDto": {
        "required": [
          "id",
          "empleadoId",
          "servicioId",
          "tipo",
          "valor",
          "activa",
          "fechaCreacionUtc",
          "fechaModificacionUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "empleadoId": {
            "type": "string",
            "format": "uuid"
          },
          "servicioId": {
            "type": "string",
            "format": "uuid"
          },
          "tipo": {
            "type": "string"
          },
          "valor": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "activa": {
            "type": "boolean"
          },
          "fechaCreacionUtc": {
            "type": "string",
            "format": "date-time"
          },
          "fechaModificacionUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}